==> exercises/11_R_functions.Rmd <==
---
title: "11_R_functions"
author: "JR"
date: "10/20/2021"
output: html_document
editor_options: 
  chunk_output_type: console
---


```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(GenomicRanges)
# source("/scratch/Shares/rinnclass/CLASS_2022/JR/CLASS_2022/util/intersect_functions.R")
# source("/scratch/Shares/rinnclass/CLASS_2022/JR/CLASS_2022/util/class_functions.R")
# source("/scratch/Shares/rinnclass/CLASS_2022/JR/CLASS_2022/util/class_functions.R")
source("/scratch/Shares/rinnclass/CLASS_2022/<your_folder>/CLASS_2022/util/intersect_functions.R")
source("/scratch/Shares/rinnclass/CLASS_2022/<your_folder./CLASS_2022/util/class_functions.R")
```

Today we will go over a "function" in R. This is a really nice feature of 
defining a function for R that you may run frequently. 

We have used many of R's built in base functions that require parameters too.
For example,table and summary we used before requires a parameter of an object to summarize etc. 
But the cool part is you can make your own functions just like these base R funcitons. 

Now, let's see what it looks like in R. A common function in life is math 
-- it takes inputs and maps it to outputs.

```{r first function}
# This function has two parameters, and we have to let R know (x & y)
# function is inside { }, then we need a 'return' to get the answer
fun <- function(x, y) {
  ans <- x * y
  return(ans)
}
#let's try it 
fun(2,-4)
# Note that the object ans doesn't exist in our environment. 
# ans is a local variable that is erased after used. global variables are in our environment.
# It's good to remember the "scope" of vars in a function don't really exist outside the function.
# Finally note a function can only return one thing!
# **** now let's make a util function *****
```

When creating a function, you can also add documentation about what that function 
does and the data types and expectation for the the parameters it takes.

```{r documenting functions}
#' A function to multiply two numbers
#'
#' @description 
#' This function will multiply the input values of X and Y
#' 
#' @param x one number you'd like to multiply
#' y the other number you'd like to multiply
fun <- function(x, y) {
  ans <- x * y
  return(ans)
}
# Let's make a util .R file and put this function in there
# clear environment and load functions -- shows up as something you can just run now
# Note that the documentation is present in the little "note" icon link.
```

Now let's understand "scope" better by creating a more complex function. 

Local variables are those variables that exist only within a certain part of a program 
like a function and are released when the function call ends.

So it's good to remember local variables are mapped to outputs through a function, but 
only exist within a function.

Let's take a look at a more complex function we may want to routinely use.

One common task is to make a list of files to be analyzed and rename them based on information
in the file name itself. Let's look at this function:

```{r base file path to peaks}
# First let's set a file path to where all our peak files are.
broadpeakfilepath <- paste0("/scratch/Shares/rinnclass/CLASS_2022/data/peaks")
# So this is now a global variable that can be called in a function.
# list.files will make a list of all the files in a dir
  fl <- list.files(broadpeakfilepath, 
                   full.names=TRUE)
  
  #or
  
  fl <- list.files("/scratch/Shares/rinnclass/CLASS_2022/data/peaks", 
                   full.names=TRUE)
  
  
  # making a list of file names as we did in 08.Rmd Note full file path is included
  # This gives us an index 1-15 where each peak file is in a vector.
  
  fl <- fl[grep("peaks.broadPeak", fl)]
  # making sure we only get files that are .broadPeak using grep
  
```

# sapply function
Let's make a function to find which DBPs have peak files. 
sapply requires a function so let's try it
```{R sapply and regular expression} 
# let's see what is in fl:
fl[1]
  tf_name <- sapply(fl, function(x){
    y <-  str_extract(x, "([^\\/]+$)")
    unlist(strsplit(y, "_"))[[1]]
    
    # We use a generic function (function(x) and set a new variable y using str_extract
    # then x is subject to str_extract and regex to split all / and words (https://regexr.com/3cr6f)  
    # strplit then splits up the broken up string from str_extract and takes the right side 
        # first character string after the last /
    
    # ?? What would the input be to make this a function ??
  })
```

# write function to import and rename a list of peak files
let's make this into a real function that is re-usable "import_peaks" and name file by DBP

```{R function to rename files by DBP}
# establishing the function using "function()"
# establishing the parameter consensus_file_path is needed
# extracting TF name as a variable along the way (check env)
import_peaks <- function(consensus_file_path = "/scratch/Shares/rinnclass/CLASS_2022/data/peaks") {
  
  # Setting some variables needed in main part of function (same as above -- peak_files & tf_name)
  peak_files <- list.files(consensus_file_path, full.names = T)
  
  # Make an object with each TF name for indexing and merging later
  tf_name <- sapply(peak_files, function(x){
    y <-  str_extract(x, "([^\\/]+$)")
    unlist(strsplit(y, "_"))[[1]]
  })
# Here is the heart of the function that will import each file as GRanges (we can use for overlaps)
# takes 
  
  # peak_list <- c()
  # for(i in 1:length(peak_files)) {
  #   # Import peak files
  #   peaks <- rtracklayer::import(peak_files[i])
  #   # Append this GRanges object to the of the list.
  #   peak_list <- c(peak_list, peaks)
  #   # Name the list elements by their TF name (we just made above)
  #   names(peak_list)[length(peak_list)] <- tf_name[i]
  # }
  peak_list = lapply(peak_files,rtracklayer::import)
  names(peak_list) = tf_name
  return(peak_list)
}
```
# ?? Why did nothing come back in the global environment ??
# running import peaks function we just made
Let's use import_peaks function but assign the local output to global output! Then recreate "peak_list"

```{R using import_peaks function -- takes a 8ish min to run}
# for didactic purposes we need to create "peak_list" to continue from the functions above 
# This is another good example of how a function only returns local vars and need to assign a global var ( <-)
# take aprox 8 min
Sys.time()
peak_list <- import_peaks("/scratch/Shares/rinnclass/CLASS_2022/data/peaks")
Sys.time()
# cool we have a list of Granges nicely named where each entry in the list is a 
# single peak file.
```
 
# Sourcing custom functions -- util folder

Let's add import peaks to util using this header (can take out all the markdown above too):

```{R sourcing custom functions from UTIL}
#' import peak .bed files as a list
#' 
#' @description 
#' this function will take each peak file and name them by the DBP
#' and return a list of GRanges peaks for each ChiPseq experiment
#' These will be used as input into create consensus peaks
#' NOTE usage of "consensus" here is bad and should just be peak file
#' **** (e.g. consensus_file_path, should be peak file path) ****
#' @param consensus_file_path the path to each peak file
```

# interesect peaks custom function
Cool now we have a function to import peak files as a GRanges.
We can now use this to find overlaps in replicates using another function
"intersect peaks" Let's take a look at our "intersect_peaks" function. 


Here we will use the peak_list we generated above and push it into this function intersect_peaks.
```{r creating intersect_peaks function}
# note the dependence on peak_list which is the output of import_peaks.
# intersect_peaks <- function(peak_list) {
#   
#   combined_peaks <- peak_list[[1]]
#   for(i in 2:length(peak_list)) {
#     suppressWarnings(pl_ov <- findOverlaps(combined_peaks, peak_list[[i]]))
#     pl1 <- combined_peaks[unique(pl_ov@from)]
#     pl2 <- peak_list[[i]][unique(pl_ov@to)]
#     suppressWarnings(combined_peaks <- GenomicRanges::reduce(union(pl1, pl2)))
#     
#   }
#   return(combined_peaks)
# }
intersect_pk = function(peaks) {
  pl_ov = GenomicRanges::findOverlaps(peaks)
  pl1 <- peaks[unique(pl_ov@from)]
  pl2 <- peaks[unique(pl_ov@to)]
  output <- GenomicRanges::reduce(union(pl1, pl2))
  return(output)
}
overlaps = lapply(peak_list,intersect_pk)
# Now that we have this function in this environment session we
# could just run it like this:
# overlaps <- intersect_peaks(peak_list)
```

```
# ***** Let's add this as a function in our util ******
```


# **** read_peaks custom function ****

The next function we need as a sub function of consensus peaks is:
read_peaks function
This function basically just filters peaks to cannonical chromosomes.

```{R read_peaks: to filter only cannonical chromosomes}
# establishing read peaks function
read_peaks <- function(broad_peak_file, filter_to_canonical_chr = TRUE) {
  dat <- read.table(broad_peak_file, sep = "\t")
  if(filter_to_canonical_chr == TRUE) {
    dat <- dat[dat$V1 %in% c(paste0("chr", 1:22), "chrM", "chrX", "chrY"),]
  }
  gr <- GRanges(seqnames = dat$V1,
                ranges = IRanges(start=dat$V2,end=dat$V3))
  return(gr)
}
# *** let's add this to our util ***
```

```{r looking deeper into read peaks}
# Let's see where the V1-V3 come from in read.table:
test <- read.table("/scratch/Shares/rinnclass/CLASS_2022/data/peaks/CTCF_R1_peaks.broadPeak")
# let's run the read_peaks function on one file
test2 <- read_peaks("/scratch/Shares/rinnclass/CLASS_2022/data/peaks/CTCF_R1_peaks.broadPeak", filter_to_canonical_chr = TRUE) 
# This is hard to read because read_peaks returns a GenomicRange
# Let's convert to a data.frame
test2 <- read_peaks("/scratch/Shares/rinnclass/CLASS_2022/data/peaks/CTCF_R1_peaks.broadPeak", filter_to_canonical_chr = TRUE) %>%
  as.data.frame()
# Compare test and test2 now -- we see 60 peaks are lost
```

The above functions are all part of another function 'create_consensus_peaks'
It is comprised of:
import_peaks
intersect_peaks 
read_peaks


# create_consensus_peaks function
Now let's look at the functions involved in creating consensus peaks.
It is also going to call other custom functions (read_peaks, intersect_peaks )

```{r create_consensus_peaks}
#' intersect replicates into a "consensus peak list" 
#' 
#' @description 
#' this function will take the  union of peak widths across replicates for a given
#' DNA binding protein. the function that will take a list of granges objects and return 
#  one granges object with merged peaks that are in all replicates
#' 
#' @param 
#'  the path to consensus peak files
#' # We're going to iterate over all the files to make it work. 
```

# Review the entire function !

Ok let's go through the code that is a function of functions we went over above.
This is the full function!

```{r reviewing the functions in the functions}
# This is just setting up to name the funciton.
create_consensus_peaks <- function(broadpeakfilepath = "/scratch/Shares/rinnclass/CLASS_2022/data/peaks/") {
  
# For now we can set broadpeakfilepath
  
# broadpeakfilepath <- "/Shares/rinn_class/data/CLASS_2022/class_exeRcises/analysis/11_consensus_peak_exercise"
# making a list of file paths to the (similar to import_peaks function)
  fl <- list.files(broadpeakfilepath, 
                   full.names=TRUE)
  fl <- fl[grep("peaks.broadPeak", fl)]
# getting a DBP name for same index as each file path
  tf_name <- sapply(fl, function(x){
    y <-  str_extract(x, "([^\\/]+$)")
    unlist(strsplit(y, "_"))[[1]]
  })
  
 
  # making sure there is a replicate and creating "unique_tf" index
  # This will be used in a forloop
  tf_df <- data.frame(table(tf_name)) %>%  # data.frame(table(tf_name))
    filter(Freq > 1)
  unique_tf <- as.character(tf_df$tf_name) # unique_tf
  
# Now a nested for loop (2 for loops) to make GRanges of peak files.
  # This is similar to read_peaks
  consensus_peaks <- list()
  for(i in 1:length(unique_tf)) {
    
    # load all the peak files corresponding to this DBP[i] in unique_tf.
    # tf <- unique_tf[1] -- allows us to look at output
    tf <- unique_tf[i]
    print(tf)
    # indexing unique DBP name to file path (e.g., first 8 are CTCF files)
    tf_index <- grep(tf, tf_name)
    # takes the TF name and grabs the index in fl for those replicates
    tf_files <- fl[tf_index]
    
    # now make a list of GRanges in a peak_list using another for loop
    # READ_PEAKS being used 
    peak_list <- c()
    for(j in 1:length(tf_files)) {
      # See the read peaks function to know what subfunctions are called.
      peak_list <- c(peak_list, read_peaks(tf_files[j]))
      # same read peaks function and we now have each DBP indexed in tf_files
    }
    # READ_PEAKS now being used
# filtering chromosomes -- redundant since read peaks does this too -- oh well.
    canonical_chr <- c(paste0("chr", 1:22), "chrM", "chrX", "chrY")
    for(i in 1:length(peak_list)) {
      peak_list[[i]] <-peak_list[[i]][which(seqnames(peak_list[[i]]) %in% canonical_chr)]
    }
# Now we use intersect_peaks functino to find overlaps 
    # INTERSECT_PEAKS now being used
    final_peakset <- intersect_peaks(peak_list = peak_list)
    if(length(final_peakset) > 0) {
      final_peakset$name <- paste0(tf, "_", 1:length(final_peakset))
    }
    
    consensus_peaks <- c(consensus_peaks, list(final_peakset))
    names(consensus_peaks)[length(consensus_peaks)] <- tf
  }
  return(consensus_peaks)
}
```


```{R running consensus peaks for test run * Somehow this only took 30 min}
# *** let's add to util and run !***
consensus_peaks <- create_consensus_peaks(broadpeakfilepath = "/scratch/Shares/rinnclass/CLASS_2022/data/peaks/") 
# export consensus peaks to CORRECT DIRECTORY!
# This takes a lot longer (~ 1.5 hrs)
for(i in 1:length(consensus_peaks)) {
  rtracklayer::export(consensus_peaks[[i]], 
                      paste0("/scratch/Shares/rinnclass/CLASS_2022/JR/CLASS_2022/class_exeRcises/analysis/11_consensus_peaks/consensus_peaks/", 
                             names(consensus_peaks)[i], 
                             "_consensus_peaks.bed"))
}
```


*********************
EXCERCISE
*********************

Make a function for formatting .broadPeak files to UCSC format.

Hint you already wrote all the code in 08 -- compile into function and put in class
functions.R in Util.

#######################################################
#######################################################
#######################################################
#######################################################
#######################################################
#######################################################
#######################################################
#######################################################

# START OF 3/7 CLASS

```{R source create_consensus_peaks}
# First make sure create_consesnus_peaks is in you Util calss_functions.R file.
# then source your util or just the functions file.
# test to make sure it's ready
# Copy lines 305-212 &  323-388 in util "class_functions.R"
source("/scratch/Shares/rinnclass/CLASS_2022/<your_folder/CLASS_2022/util/class_functions.R")
```

# Now we are going to create an rscript to run create_consensus_peaks
# **** Must have output folder ready ****
# *** Set working directory to 11_ , create necessary directories ***
```{r create R script}
# the rscript is going to run somewhere else in a blank environment.
# so we need to have everything we need to run in the rscript
library(tidyverse)
library(GenomicRanges)
# source("/scratch/Shares/rinnclass/CLASS_2022/JR/CLASS_2022/util/intersect_functions.R")
source("/scratch/Shares/rinnclass/CLASS_2022/JR/CLASS_2022/util/class_functions.R")
# run consensus peaks
consensus_peaks <- create_consensus_peaks("/scratch/Shares/rinnclass/CLASS_2022/data/peaks")
# export consensus peaks 
for(i in 1:length(consensus_peaks)) {
  rtracklayer::export(consensus_peaks[[i]], 
                      paste0("/scratch/Shares/rinnclass/CLASS_2022/JR/CLASS_2022/class_exeRcises/analysis/11_consensus_peaks/consensus_peaks/", 
                             names(consensus_peaks)[i], 
                             "_consensus_peaks.bed"))
}
# run in console pane under "jobs" select Rscript from /analysis/11_consensus_peaks
```

# While create_consensus_peaks is running let's dissect the function with a single example.

********************
TEST EXAMPLE : CTCF under the hood of create_consensus_peaks
********************


```{R test example of create_consensus_peaks}
# This how we would now run the function for all DBPs -- however let's step through one example
# source("/scratch/Shares/rinnclass/CLASS_2022/JR/CLASS_2022/util/class_functions.R")
source("/scratch/Shares/rinnclass/CLASS_2022/<your_folder>/CLASS_2022/util/class_functions.R")
# create_consensus_peaks <- function(broadpeakfilepath = "/scratch/Shares/rinnclass/CLASS_2022/data/peaks/")
fl <- list.files("/scratch/Shares/rinnclass/CLASS_2022/data/peaks", 
                 full.names=TRUE)
# let's take a look -- a bunch of file paths
fl <- fl[grep("peaks.broadPeak", fl)]
# now let's look (nothing changed since all are broad_peaks)
fl
# getting a DBP name for same index as each file path
tf_name <- sapply(fl, function(x){
  y <-  str_extract(x, "([^\\/]+$)")
  unlist(strsplit(y, "_"))[[1]]
})
# let's take a look
tf_name
# holds all the names extracted above from the file name
# making sure there is a replicate and creating "unique_tf" index
# We could do this a few ways
table(table(tf_name))
# we see that three DBPs only have one replicate !!
# we could make a dataframe as is done in create_consensus peaks
tf_df <- data.frame(table(tf_name)) %>%  # data.frame(table(tf_name))
  filter(Freq > 1)
unique_tf <- as.character(tf_df$tf_name) # unique_tf
# Now a nested for loop (2 for loops) to make GRanges of peak files.
# This is similar to read_peaks
consensus_peaks <- list()
#taking out the for loop since we are only doing one example:
# for(i in 1:length(unique_tf)) {
tf <- unique_tf[37]
print(tf)
# indexing unique DBP name to file path (e.g., first 8 are CTCF files)
tf_index <- grep(tf, tf_name)
# takes the TF name and grabs the index in fl for those replicates
tf_index
# now we are subsetting tf_files to only CTCF indices
tf_files <- fl[tf_index]
tf_files
# now make a list of GRanges in a peak_list using another for loop
# READ_PEAKS being used 
peak_list <- c()
# taking out the for loop for one example
for(j in 1:length(tf_files)) {
  # Using read peaks to clean up chromosomes
  peak_list <- c(peak_list, read_peaks(tf_files[j]))
  # note this is concatinating each DBP when run in for loop (the next j will be pasted into peak_list)
  # peak_list <-  read_peaks(tf_files[[j]], filter_to_canonical_chr = TRUE)
  # same read peaks function and we now have each DBP indexed in tf_files
}
# READ_PEAKS now being used
# filtering chromosomes -- redundant since read peaks does this too -- oh well.
canonical_chr <- c(paste0("chr", 1:22), "chrM", "chrX", "chrY")
for(i in 1:length(peak_list)) {
  peak_list[[i]] <-peak_list[[i]][which(seqnames(peak_list[[i]]) %in% canonical_chr)]
  
}
# sapply(peak_list, length)
# Now we use intersect_peaks functino to find overlaps 
# INTERSECT_PEAKS now being used
final_peakset <- intersect_peaks(peak_list = peak_list)
# this is findOverlaps for file 1-2, then that file is used for 3
if(length(final_peakset) > 0) {
  final_peakset$name <- paste0(tf, "_", 1:length(final_peakset))
}
# concatinating to add the next DBP to this list of consensus_peaks
consensus_peaks <- c(consensus_peaks, list(final_peakset))
names(consensus_peaks)[length(consensus_peaks)] <- tf
```


*********************
EXCERCISE
*********************

Make a function for formatting .broadPeak files to UCSC format.

Hint you already wrote all the code in 08 -- compile into function and put in class
functions.R in Util.
